+++ lib/rule.ml
@@ -2434,8 +2434,10 @@ let extract_style_with_rules ~sel ~class_name ?merge_key ~props rule_list =
 let outputs ?(theme = Scheme.default) ?order_tbl util =
   let rec utility_order = function
     | Utility.Base b -> Some (Utility.order b)
-    | Utility.Modified (_, u) | Utility.Important (_, u) | Utility.Aliased (_, u)
-      ->
+    | Utility.Modified (_, u)
+    | Utility.Important (_, u)
+    | Utility.Aliased (_, u)
+    | Utility.Theme_bound (_, u) ->
         utility_order u
     | Utility.Group _ -> None
   in